Move event API parts out of internal packages
