+++ silk-commands/src/main/kotlin/net/silkmc/silk/commands/event/CommandEvents.kt
@@ -1,4 +1,4 @@
-package net.silkmc.silk.commands.internal.events
+package net.silkmc.silk.commands.event
 
 import com.mojang.brigadier.CommandDispatcher
 import net.minecraft.commands.CommandBuildContext
@@ -6,8 +6,8 @@ import net.minecraft.commands.CommandSourceStack
 import net.minecraft.commands.Commands.CommandSelection
 import net.silkmc.silk.commands.ClientCommandSourceStack
 import net.silkmc.silk.core.annotations.ExperimentalSilkApi
-import net.silkmc.silk.core.events.Events
-import net.silkmc.silk.core.internal.events.Event
+import net.silkmc.silk.core.event.Event
+import net.silkmc.silk.core.event.Events
 
 /**
  * Events related to commands.
@@ -21,7 +21,7 @@ object CommandEvents {
     class RegisterEvent(
         val dispatcher: CommandDispatcher<CommandSourceStack>,
         val context: CommandBuildContext,
-        val selection: CommandSelection
+        val selection: CommandSelection,
     )
 
     class RegisterEventClient(
