--- silk-core/src/main/kotlin/net/silkmc/silk/core/events/Events.kt
@@ -1,10 +1,0 @@
-package net.silkmc.silk.core.events
-
-import net.silkmc.silk.core.annotations.ExperimentalSilkApi
-
-/**
- * The base object for discovering events, check the extensions
- * values of this object.
- */
-@ExperimentalSilkApi
-object Events
