+++ silk-core/src/main/kotlin/net/silkmc/silk/core/internal/SilkEntrypoint.kt
@@ -4,8 +4,8 @@ import net.fabricmc.api.ClientModInitializer
 import net.fabricmc.api.ModInitializer
 import net.silkmc.silk.core.Silk
 import net.silkmc.silk.core.annotations.InternalSilkApi
-import net.silkmc.silk.core.events.Events
-import net.silkmc.silk.core.internal.events.server
+import net.silkmc.silk.core.event.Events
+import net.silkmc.silk.core.event.server
 import net.silkmc.silk.core.logging.logInfo
 import net.silkmc.silk.core.task.LifecycleTasksManager
 import net.silkmc.silk.core.task.McCoroutineManager
