+++ silk-core/src/main/kotlin/net/silkmc/silk/core/task/McCoroutineManager.kt
@@ -3,14 +3,15 @@ package net.silkmc.silk.core.task
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.SupervisorJob
 import kotlinx.coroutines.asCoroutineDispatcher
-import net.silkmc.silk.core.internal.events.ServerEvents
+import net.silkmc.silk.core.event.Events
+import net.silkmc.silk.core.event.server
 import net.silkmc.silk.core.logging.logger
 
 internal object McCoroutineManager {
     private val log = logger()
 
     fun init() {
-        ServerEvents.Init.register {
+        Events.server.preStart.listen {
             mcCoroutineDispatcher = it.server.asCoroutineDispatcher()
             mcCoroutineScope = CoroutineScope(SupervisorJob() + mcCoroutineDispatcher)
             log.info("Initialized mcCoroutineScope (MinecraftServer as executor)")
